Tweets airflow loading/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages

pandas
tweepy
regex
nltk
snowflake-connector-python
apache-airflow-providers-snowflake